requirements.txt
ctransformers==0.2.5 
sentence-transformers==2.2.2
pinecone-client
flask
pypdf
langchain
pandas
numpy
python-dotenv

-e .